# docs/api/openapi.json
{"openapi":"3.0.0","paths":{"/api/v1/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"summary":"Check service health","tags":["Health"]}},"/api/v1/health/live":{"get":{"operationId":"HealthController_live","parameters":[],"responses":{"200":{"description":""}},"summary":"Liveness probe","tags":["Health"]}},"/api/v1/health/ready":{"get":{"operationId":"HealthController_ready","parameters":[],"responses":{"200":{"description":""}},"summary":"Readiness probe","tags":["Health"]}},"/api/v1/metrics":{"get":{"operationId":"PrometheusController_index","parameters":[],"responses":{"200":{"description":""}},"tags":["Prometheus"]}},"/api/v1/upload":{"post":{"operationId":"UploadController_uploadFile","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Upload a document file","tags":["Upload"]}},"/api/v1/documents":{"post":{"operationId":"DocumentsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDocumentDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Upload a document","tags":["Documents"]},"get":{"operationId":"DocumentsController_findAll","parameters":[{"name":"status","required":true,"in":"query","schema":{"type":"string"}},{"name":"source","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List documents","tags":["Documents"]}},"/api/v1/documents/{id}":{"get":{"operationId":"DocumentsController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get document by ID","tags":["Documents"]},"delete":{"operationId":"DocumentsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Soft delete document","tags":["Documents"]}},"/api/v1/documents/{id}/process":{"post":{"operationId":"DocumentsController_process","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Trigger document processing","tags":["Documents"]}},"/api/v1/search":{"get":{"operationId":"SearchController_search","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}},{"name":"mode","required":true,"in":"query","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Hybrid enterprise search","tags":["Search"]}},"/api/v1/search/suggestions":{"get":{"operationId":"SearchController_suggestions","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get search suggestions","tags":["Search"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Returns JWT tokens"}},"summary":"Authenticate user","tags":["Authentication"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"summary":"Register new user","tags":["Authentication"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":""}},"summary":"Refresh access token","tags":["Authentication"]}},"/api/v1/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"Returns the authenticated user"}},"security":[{"bearer":[]}],"summary":"Get current user profile","tags":["Authentication"]}},"/api/v1/connectors":{"post":{"operationId":"ConnectorsController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create a new connector","tags":["Connectors"]},"get":{"operationId":"ConnectorsController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List connectors","tags":["Connectors"]}},"/api/v1/connectors/{id}":{"get":{"operationId":"ConnectorsController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get connector details","tags":["Connectors"]},"put":{"operationId":"ConnectorsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update connector","tags":["Connectors"]},"delete":{"operationId":"ConnectorsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete connector","tags":["Connectors"]}},"/api/v1/connectors/{id}/sync":{"post":{"operationId":"ConnectorsController_sync","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Trigger connector sync","tags":["Connectors"]}},"/api/v1/connectors/{id}/runs":{"get":{"operationId":"ConnectorsController_getRuns","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get sync run history","tags":["Connectors"]}},"/api/v1/graph/nodes":{"get":{"operationId":"GraphController_getNodes","parameters":[{"name":"type","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}},{"name":"skip","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Get graph nodes","tags":["Knowledge Graph"]}},"/api/v1/graph/nodes/{id}":{"get":{"operationId":"GraphController_getNode","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get node by ID with relationships","tags":["Knowledge Graph"]}},"/api/v1/graph/search":{"get":{"operationId":"GraphController_searchNodes","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Search nodes","tags":["Knowledge Graph"]}},"/api/v1/graph/subgraph/{id}":{"get":{"operationId":"GraphController_getSubgraph","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"depth","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"Get subgraph around a node","tags":["Knowledge Graph"]}},"/api/v1/graph/query":{"post":{"operationId":"GraphController_query","parameters":[],"responses":{"201":{"description":""}},"summary":"Execute raw Cypher query","tags":["Knowledge Graph"]}},"/api/v1/chat/messages":{"post":{"operationId":"ChatController_sendMessage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Send a message and get AI response","tags":["Chat"]}},"/api/v1/chat/conversations":{"get":{"operationId":"ChatController_listConversations","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List user conversations","tags":["Chat"]}},"/api/v1/chat/conversations/{id}":{"get":{"operationId":"ChatController_getConversation","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get conversation history","tags":["Chat"]}},"/api/v1/meetings":{"post":{"operationId":"MeetingsController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create a meeting record","tags":["Meetings"]},"get":{"operationId":"MeetingsController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List meetings","tags":["Meetings"]}},"/api/v1/meetings/{id}":{"get":{"operationId":"MeetingsController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get meeting details","tags":["Meetings"]},"delete":{"operationId":"MeetingsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete meeting","tags":["Meetings"]}},"/api/v1/meetings/{id}/summarize":{"post":{"operationId":"MeetingsController_summarize","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Generate AI meeting summary","tags":["Meetings"]}},"/api/v1/notifications":{"get":{"operationId":"NotificationsController_findAll","parameters":[{"name":"unreadOnly","required":true,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List notifications","tags":["Notifications"]}},"/api/v1/notifications/unread-count":{"get":{"operationId":"NotificationsController_getUnreadCount","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get unread notification count","tags":["Notifications"]}},"/api/v1/notifications/{id}/read":{"post":{"operationId":"NotificationsController_markAsRead","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Mark notification as read","tags":["Notifications"]}},"/api/v1/notifications/read-all":{"post":{"operationId":"NotificationsController_markAllAsRead","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Mark all notifications as read","tags":["Notifications"]}},"/api/v1/notifications/{id}":{"delete":{"operationId":"NotificationsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete notification","tags":["Notifications"]}},"/api/v1/admin/dashboard":{"get":{"operationId":"AdminController_getDashboard","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get organization dashboard stats","tags":["Admin"]}},"/api/v1/admin/audit-logs":{"get":{"operationId":"AdminController_getAuditLogs","parameters":[{"name":"entity","required":true,"in":"query","schema":{"type":"string"}},{"name":"action","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get audit logs","tags":["Admin"]}},"/api/v1/admin/health":{"get":{"operationId":"AdminController_getHealth","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get system health","tags":["Admin"]}},"/api/v1/policies":{"post":{"operationId":"PoliciesController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create a policy","tags":["Policies"]},"get":{"operationId":"PoliciesController_findAll","parameters":[{"name":"category","required":true,"in":"query","schema":{"type":"string"}},{"name":"active","required":true,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List policies","tags":["Policies"]}},"/api/v1/policies/search":{"get":{"operationId":"PoliciesController_search","parameters":[{"name":"q","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Search policies","tags":["Policies"]}},"/api/v1/policies/{id}":{"get":{"operationId":"PoliciesController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get policy details","tags":["Policies"]},"put":{"operationId":"PoliciesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update policy","tags":["Policies"]},"delete":{"operationId":"PoliciesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Delete policy","tags":["Policies"]}},"/api/v1/expertise/search":{"get":{"operationId":"ExpertiseController_findExperts","parameters":[{"name":"topic","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Find experts by topic","tags":["Expertise"]}},"/api/v1/expertise/summary":{"get":{"operationId":"ExpertiseController_getSummary","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get organization expertise summary","tags":["Expertise"]}},"/api/v1/gaps":{"get":{"operationId":"GapsController_getGaps","parameters":[{"name":"severity","required":true,"in":"query","schema":{"type":"string"}},{"name":"category","required":true,"in":"query","schema":{"type":"string"}},{"name":"resolved","required":true,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List detected knowledge gaps","tags":["Knowledge Gaps"]}},"/api/v1/gaps/detect":{"post":{"operationId":"GapsController_detect","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Run knowledge gap detection","tags":["Knowledge Gaps"]}},"/api/v1/gaps/{id}/resolve":{"post":{"operationId":"GapsController_resolve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Mark gap as resolved","tags":["Knowledge Gaps"]}},"/api/v1/recommendations":{"get":{"operationId":"RecommendationsController_getRecommendations","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get personalized recommendations","tags":["Recommendations"]}},"/api/v1/recommendations/feed":{"get":{"operationId":"RecommendationsController_getFeed","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get personalized feed","tags":["Recommendations"]}}},"info":{"title":"AI Knowledge Graph API","description":"Enterprise AI-powered Organizational Knowledge Engine","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"api_key":{"type":"apiKey","in":"header","name":"api_key"}},"schemas":{"CreateDocumentDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"filePath":{"type":"string"},"fileType":{"type":"string"},"fileSize":{"type":"number"},"mimeType":{"type":"string"},"checksum":{"type":"string"},"source":{"type":"string","enum":["UPLOAD","GOOGLE_DRIVE","ONEDRIVE","SHAREPOINT","SLACK","TEAMS","NOTION","CONFLUENCE","GITHUB","GITLAB","JIRA","LINEAR","DROPBOX","EMAIL","MEETING"]},"sourceUrl":{"type":"string"},"workspaceId":{"type":"string"}},"required":["title","filePath","fileType","fileSize","mimeType","checksum"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@company.com"},"password":{"type":"string","example":"securePassword123"}},"required":["email","password"]},"RegisterDto":{"type":"object","properties":{"email":{"type":"string","example":"john@company.com"},"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"password":{"type":"string","example":"securePassword123"},"organizationName":{"type":"string"}},"required":["email","firstName","lastName","password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"SendMessageDto":{"type":"object","properties":{"content":{"type":"string","example":"How do we deploy the Payment API?"},"conversationId":{"type":"string"}},"required":["content"]}}}}